+++ k6/script.k8s.js
@@ -1,0 +1,25 @@
+import http from 'k6/http';
+import { check } from 'k6';
+
+export const options = {
+  vus: 20,
+  duration: '10s',
+};
+
+export default function () {
+  const url = 'http://192.168.49.2/api/v1/reserveTable';
+  const payload = JSON.stringify({
+    "numberOfCustomers": 3
+  });
+
+  const params = {
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  };
+
+  const res = http.post(url, payload, params);
+  check(res, {
+    'is status 200': (r) => r.status === 200,
+  });
+}
